docs: crypto
