+++ crates/outlayer/sdk/src/host/crypto/secp256k1.rs
@@ -3,9 +3,23 @@ use defuse_outlayer_host as host;
 #[cfg(target_family = "wasm")]
 use defuse_outlayer_sys as sys;
 
+/// Secp256k1 (a.k.a. k256) public key in SEC-1 uncompressed form
+/// **without** leading tag byte (0x04)
 pub type PublicKey = [u8; 64];
+
+/// Secp256k1 (a.k.a k256) signature encoded as concatenated `r`, `s` and
+/// `v` (recovery byte)
 pub type Signature = [u8; 65];
 
+/// Derive public key from root for given application-specific path.
+///
+/// The derivation is **non-hierarchical** (or "plain"): derived
+/// keys **do not** form a tree-like structure. Instead, child keys
+/// are all derived from a single root key and can be considered as
+/// "peers" to each other.
+///
+/// Returns secp256k1 public key encoded in SEC-1 uncompressed form
+/// **without** leading tag byte (0x04).
 pub fn derive_public_key(path: impl AsRef<str>) -> PublicKey {
     #[cfg(target_family = "wasm")]
     {
@@ -24,6 +38,10 @@ pub fn derive_public_key(path: impl AsRef<str>) -> PublicKey {
     .expect("invalid length")
 }
 
+/// Sign given message with a secret key **intetnally** derived for
+/// given application-specific path.
+///
+/// Returns a signature as concatenated `r`, `s` and `v` (recovery byte).
 pub fn sign(path: impl AsRef<str>, msg: impl AsRef<[u8]>) -> Signature {
     #[cfg(target_family = "wasm")]
     {
