+++ crates/outlayer/wit/deps/crypto.wit
@@ -1,12 +1,37 @@
 package outlayer:crypto@0.1.0;
 
+/// Ed25519
 interface ed25519 {
+    /// Derive public key from root for given application-specific path.
+    ///
+    /// The derivation is **non-hierarchical** (or "plain"): derived
+    /// keys **do not** form a tree-like structure. Instead, child keys
+    /// are all derived from a single root key and can be considered as
+    /// "peers" to each other.
     derive-public-key: func(path: string) -> list<u8>;
+
+    /// Sign given message with a secret key **intetnally** derived for
+    /// given application-specific path.
     sign: func(path: string, msg: list<u8>) -> list<u8>;
 }
 
+/// Secp256k1 (a.k.a. k256)
 interface secp256k1 {
+    /// Derive public key from root for given application-specific path.
+    ///
+    /// The derivation is **non-hierarchical** (or "plain"): derived
+    /// keys **do not** form a tree-like structure. Instead, child keys
+    /// are all derived from a single root key and can be considered as
+    /// "peers" to each other.
+    ///
+    /// Returns secp256k1 public key encoded in SEC-1 uncompressed form
+    /// **without** leading tag byte (0x04).
     derive-public-key: func(path: string) -> list<u8>;
+
+    /// Sign given message with a secret key **intetnally** derived for
+    /// given application-specific path.
+    ///
+    /// Returns a signature as concatenated `r`, `s` and `v` (recovery byte).
     sign: func(path: string, msg: list<u8>) -> list<u8>;
 }
